requirements.txt
amqp==5.1.1
anyio==3.6.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
fastapi==0.87.0
h11==0.14.0
idna==3.4
kombu==5.2.4
packaging==21.3
prompt-toolkit==3.0.32
psycopg2==2.9.5
pydantic==1.10.2
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.6
redis==4.3.4
requests==2.28.1
six==1.16.0
sniffio==1.3.0
starlette==0.21.0
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
